add partial class containing convenience methods
